Start the aarch64 port on the BSDs that are not macOS
os/bsd/os_bsd.cpp had no AARCH64 row in the table dll_load uses to check
that a library matches the running machine, so it stopped at the #error
listing the architectures it knows.

The signal context is where the four systems differ most: Darwin keeps
the registers behind a pointer with UNIX03 member prefixes, FreeBSD in
mc_gpregs, NetBSD in a __gregs array indexed by <machine/reg.h>'s names,
and OpenBSD straight in the struct sigcontext the handler is given.  Give
each its own spelling of the context_ macros the rest of the file uses.

HotSpot cross-modifies code, and on Darwin/aarch64 it leans on
pthread_jit_write_protect_np for the barriers that makes safe.  The other
BSDs have no such call, and the Cortex-specific workarounds used
elsewhere are not enough on Apple Silicon, so issue the DSB and ISB
directly.  That is what ThreadWXEnable now does under _BSDONLY_SOURCE,
and the guard around its uses widens from MACOS_AARCH64_ONLY to
BSD_AARCH64_ONLY.

A pointer difference is ptrdiff_t, which is long.  Where int64_t is long
long -- OpenBSD spells it that way -- that matches neither g_uabs(jlong)
nor g_uabs(julong), so the branch-range check has to say which it means.

+++ src/hotspot/cpu/aarch64/assembler_aarch64.hpp
@@ -930,7 +930,10 @@ public:
   static const uint64_t branch_range = NOT_DEBUG(128 * M) DEBUG_ONLY(2 * M);
 
   static bool reachable_from_branch_at(address branch, address target) {
-    return g_uabs(target - branch) < branch_range;
+    // The difference of two addresses is ptrdiff_t, which is long.  Where
+    // int64_t is long long -- OpenBSD spells it that way -- that matches
+    // neither g_uabs(jlong) nor g_uabs(julong) exactly, so say which.
+    return g_uabs((jlong)(target - branch)) < branch_range;
   }
 
   // Unconditional branch (immediate)
